+++ libs/core/src/state/settings/by_monitor.rs
@@ -4,12 +4,12 @@ use std::collections::HashMap;
 
 use crate::{
     resource::WidgetId,
-    state::{by_widget::ThirdPartyWidgetSettings, WorkspaceId},
+    state::{by_widget::GenericWidgetSettings, WorkspaceId},
 };
 
 #[derive(Debug, Default, Clone, Serialize, Deserialize, JsonSchema)]
 #[cfg_attr(all(feature = "gen-binds", not(feature = "salvo")), derive(ts_rs::TS))]
-pub struct MonitorSettingsByWidget(HashMap<WidgetId, ThirdPartyWidgetSettings>);
+pub struct MonitorSettingsByWidget(HashMap<WidgetId, GenericWidgetSettings>);
 
 impl MonitorSettingsByWidget {
     pub fn is_widget_enabled(&self, widget_id: &WidgetId) -> bool {
@@ -18,11 +18,11 @@ impl MonitorSettingsByWidget {
             .is_none_or(|settings| settings.enabled)
     }
 
-    pub fn remove(&mut self, widget_id: &WidgetId) -> Option<ThirdPartyWidgetSettings> {
+    pub fn remove(&mut self, widget_id: &WidgetId) -> Option<GenericWidgetSettings> {
         self.0.remove(widget_id)
     }
 
-    pub fn insert(&mut self, widget_id: WidgetId, settings: ThirdPartyWidgetSettings) {
+    pub fn insert(&mut self, widget_id: WidgetId, settings: GenericWidgetSettings) {
         self.0.insert(widget_id, settings);
     }
 }